+++ CHANGELOG.md
@@ -133,6 +133,7 @@ v0.19
 * Experimental support for ``ResourceFixturesMixin`` that can be used to load resource content through fixture files.
 * Cancelling a ``HttpFileResource`` will result in an empty body instead of a body of None.
 * Removes ``load_config`` defaults parameter.
+* Adds optional defaults to ``reach`` function when path can't be found in the given data structure.
 
 
 v0.18